handle no idp
