+++ fence/blueprints/data/indexd.py
@@ -67,9 +67,9 @@ def get_signed_url_for_file(
 ):
     requested_protocol = requested_protocol or flask.request.args.get("protocol", None)
     r_pays_project = flask.request.args.get("userProject", None)
+
     # default to signing the url even if it's a public object
     # this will work so long as we're provided a user token
-
     force_signed_url = True
     no_force_sign_param = flask.request.args.get("no_force_sign")
     if no_force_sign_param and no_force_sign_param.lower() == "true":
